# Fix performance of adding SecondOrderCone constraints